Implemented view/edit modes
