+++ client/app/assets/less/ant.less
@@ -362,6 +362,11 @@
       border-radius: 50%;
     }
   }
+
+  &.form-item-line-height-normal .@{form-prefix-cls}-item-control {
+    line-height: 20px;
+    margin-top: 9px;
+  }
 }
 
 // TODO: width vars
