+++ client/app/pages/alert/components/Query.less
@@ -1,0 +1,7 @@
+.alert-query-link {
+  font-weight: 600;
+  text-decoration: underline;
+  display: inline-block;
+  position: relative;
+  top: -1px;
+}
