+++ client/app/pages/alert/components/Rearm.less
@@ -1,0 +1,8 @@
+.alert-rearm > * {
+  vertical-align: top;
+  margin-right: 8px !important;
+
+  &.ant-select {
+      width: auto !important;
+  }
+}
